requirements.txt
click==8.1.3
cryptography==39.0.2
dataclasses-json==0.5.7
Flask==2.2.3
Flask-Login==0.6.2
gpt-index==0.4.40
gunicorn==20.1.0
Jinja2==3.1.2
langchain==0.0.123
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.0
pandas==1.5.3
pyOpenSSL==23.0.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==1.4.47
urllib3==1.26.14
Werkzeug==2.2.3
